Code reformatted and comments added
